Task 5.2 - box & sphere, no tests

+++ framework/shape.hpp
@@ -7,8 +7,8 @@ public:
     Shape();
 
     //functions
-    virtual float area() const = 0;
-	virtual float volume() const = 0;
+    virtual double area() const = 0;
+	virtual double volume() const = 0;
 };
 
 #endif
